refactor: move IPC message classes under `IpcServer` proj
