+++ GlazeWM.Application.IpcServer/Messages/GetMonitorsMessage.cs
@@ -1,6 +1,6 @@
 using CommandLine;
 
-namespace GlazeWM.Domain.Common.IpcMessages
+namespace GlazeWM.Application.IpcServer.Messages
 {
   [Verb("monitors", HelpText = "Get all monitors.")]
   public class GetMonitorsMessage
