+++ GlazeWM.Application.IpcServer/Messages/GetWindowsMessage.cs
@@ -1,6 +1,6 @@
 using CommandLine;
 
-namespace GlazeWM.Domain.Common.IpcMessages
+namespace GlazeWM.Application.IpcServer.Messages
 {
   [Verb("windows", HelpText = "Get all windows.")]
   public class GetWindowsMessage
